Implementation of individuals as Accounts.

+++ src/java/org/aspcfs/modules/accounts/actions/Contacts.java
@@ -186,10 +186,16 @@ public final class Contacts extends CFSModule {
       htmlDialog.setRelationships(thisContact.processDependencies(db));
 
       if (!thisContact.hasAccount()) {
-        htmlDialog.setTitle("CFS: Confirm Delete");
-        htmlDialog.setHeader("The contact you are requesting to delete has the following dependencies within CFS:");
-        htmlDialog.addButton("Delete All", "javascript:window.location.href='Contacts.do?command=Delete&orgId=" + orgId + "&id=" + id + "'");
-        htmlDialog.addButton("Cancel", "javascript:parent.window.close()");
+        
+        if (thisContact.getPrimaryContact()) {
+          htmlDialog.setHeader("This contact cannot be deleted because it is also an individual account.");
+          htmlDialog.addButton("OK", "javascript:parent.window.close()");
+        } else {
+          htmlDialog.setTitle("CFS: Confirm Delete");
+          htmlDialog.setHeader("The contact you are requesting to delete has the following dependencies within CFS:");
+          htmlDialog.addButton("Delete All", "javascript:window.location.href='Contacts.do?command=Delete&orgId=" + orgId + "&id=" + id + "'");
+          htmlDialog.addButton("Cancel", "javascript:parent.window.close()");
+        }
       } else {
         htmlDialog.setHeader("This contact cannot be deleted because it is associated with a User account.");
         htmlDialog.addButton("OK", "javascript:parent.window.close()");
@@ -382,6 +388,13 @@ public final class Contacts extends CFSModule {
 
     Contact newContact = (Contact)context.getFormBean();
     newContact.setRequestItems(context.getRequest());
+    
+    if (context.getRequest().getParameter("primaryContact") != null) {
+      if (context.getRequest().getParameter("primaryContact").equalsIgnoreCase("true")) {
+        newContact.setPrimaryContact(true);
+      }
+    }
+    
     Organization thisOrganization = null;
     String orgid = context.getRequest().getParameter("orgId");
